Add validation of short_type again
This switches the old validation code for short_type to work in the
concern. Also updates the spec to check if a string shorter than the max
is also valid.
